+++ k8s/src/main/scala/io/buoyant/k8s/AuthFilter.scala
@@ -1,7 +1,7 @@
 package io.buoyant.k8s
 
-import com.twitter.finagle.{Service, SimpleFilter}
 import com.twitter.finagle.http.{Request, Response}
+import com.twitter.finagle.{Service, SimpleFilter}
 
 class AuthFilter(token: String) extends SimpleFilter[Request, Response] {
   def apply(req: Request, service: Service[Request, Response]) = {