+++ namer/consul/src/main/scala/io/buoyant/namer/consul/ConsulInitializer.scala
@@ -4,7 +4,8 @@ import com.fasterxml.jackson.annotation.JsonIgnore
 import com.twitter.finagle.http.{Request, Response}
 import com.twitter.finagle.param.Label
 import com.twitter.finagle.tracing.NullTracer
-import com.twitter.finagle.{Filter, Http, Path, Stack}
+import com.twitter.finagle._
+import com.twitter.util.Monitor
 import io.buoyant.config.types.Port
 import io.buoyant.consul.{SetAuthTokenFilter, SetHostFilter, v1}
 import io.buoyant.namer.{NamerConfig, NamerInitializer}
@@ -19,6 +20,7 @@ import io.buoyant.namer.{NamerConfig, NamerInitializer}
  *   host: consul.site.biz
  *   port: 8600
  *   includeTag: true
+ *   useHealthCheck: false
  *   setHost: true
  *   token: some-consul-acl-token
  * </pre>
@@ -34,6 +36,7 @@ case class ConsulConfig(
   host: Option[String],
   port: Option[Port],
   includeTag: Option[Boolean],
+  useHealthCheck: Option[Boolean],
   token: Option[String] = None,
   setHost: Option[Boolean] = None
 ) extends NamerConfig {
@@ -54,23 +57,27 @@ case class ConsulConfig(
    * Build a Namer backed by Consul.
    */
   @JsonIgnore
-  def newNamer(params: Stack.Params): CatalogNamer = {
+  def newNamer(params: Stack.Params): ConsulNamer = {
     val authFilter = token match {
       case Some(t) => new SetAuthTokenFilter(t)
       case None => Filter.identity[Request, Response]
     }
     val filters = new SetHostFilter(getHost, getPort) andThen authFilter
+    val interruptionMonitor = Monitor.mk {
+      case e: Failure if e.isFlagged(Failure.Interrupted) => true
+    }
 
     val service = Http.client
+      .withMonitor(interruptionMonitor)
       .withParams(Http.client.params ++ params)
       .configured(Label("namer" + prefix))
       .withTracer(NullTracer)
       .filtered(filters)
       .newService(s"/$$/inet/$getHost/$getPort")
 
-    new CatalogNamer(
+    new ConsulNamer(
       prefix,
-      v1.CatalogApi(service),
+      if (useHealthCheck.getOrElse(false)) v1.HealthApi(service) else v1.CatalogApi(service),
       v1.AgentApi(service),
       includeTag = includeTag.getOrElse(false),
       setHost = setHost.getOrElse(false)