+++ router/http/src/test/scala/io/buoyant/router/http/ProxyRewriteFiltertest.scala
@@ -6,7 +6,7 @@ import com.twitter.util.Future
 import io.buoyant.test.Awaits
 import org.scalatest.FunSuite
 
-class DiscardProxySideEffectsTest extends FunSuite with Awaits {
+class ProxyRewriteFilterTest extends FunSuite with Awaits {
 
   def service: Service[Request, Response] = {
     val svc = Service.mk[Request, Response] { req =>
@@ -17,15 +17,15 @@ class DiscardProxySideEffectsTest extends FunSuite with Awaits {
       Future.value(rsp)
     }
 
-    val stk = DiscardProxySideEffects.module.toStack(
-      Stack.Leaf(DiscardProxySideEffects.module.role, ServiceFactory.const(svc))
+    val stk = ProxyRewriteFilter.module.toStack(
+      Stack.Leaf(ProxyRewriteFilter.module.role, ServiceFactory.const(svc))
     )
 
     await(stk.make(Stack.Params.empty)())
   }
 
-  test("filter rewrites uri and drops Proxy-* headers for proxied requests") {
-    val req = Request(Method.Get, "http://acme.co:8080/foo;matrix=uri;/bar?x=1&y=2")
+  test("rewrite proxy requests as non-proxy requests") {
+    val req = Request(Method.Get, "http://acme.co:8080/foo;matrix=uri;/bar?x=1&y=2&z=%28%20hi%20%29")
     req.host = "example.com:9000"
     req.headerMap.set("Proxy-Connection", "Keep-Alive")
     req.headerMap.set("Proxy-Authorization", "Basic QWxhZGRpbjpvcGVuIHNlc2FtZQ==")
@@ -34,16 +34,15 @@ class DiscardProxySideEffectsTest extends FunSuite with Awaits {
     req.headerMap.set("Pragma", "42")
 
     val result = await(service(req))
-    assert(result.getContentString() == "/foo;matrix=uri;/bar?x=1&y=2")
-    assert(
-      result.headerMap.toSet == Set(
-        ("Host", "acme.co:8080"),
-        ("Pragma", "42")
-      )
-    )
+    assert(result.getContentString() == "/foo;matrix=uri;/bar?x=1&y=2&z=(%20hi%20)")
+    assert(result.headerMap == Map(
+      "Host" -> "acme.co:8080",
+      "Pragma" -> "42",
+      "Proxy-Foo" -> "Bar"
+    ))
   }
 
-  test("filter drop Proxy-* Headers even for requests that doesn't look like proxied") {
+  test("strip Proxy-* Headers for non-proxy requests") {
     val req = Request(Method.Get, "/foo;matrix=uri;/bar?x=1&y=2")
     req.host = "acme.co:8080"
     req.headerMap.set("Proxy-Connection", "Keep-Alive")
@@ -54,11 +53,10 @@ class DiscardProxySideEffectsTest extends FunSuite with Awaits {
 
     val result = await(service(req))
     assert(result.getContentString() == "/foo;matrix=uri;/bar?x=1&y=2")
-    assert(
-      result.headerMap.toSet == Set(
-        ("Host", "acme.co:8080"),
-        ("Pragma", "42")
-      )
-    )
+    assert(result.headerMap == Map(
+      "Host" -> "acme.co:8080",
+      "Pragma" -> "42",
+      "Proxy-Foo" -> "Bar"
+    ))
   }
 }